Basic ServiceNow integration working to retrieve incidents.
The integration retrieve incidents, I left some TODO comments since I want to refine it further to handle issues with queries that are failing and need to be attempted again.

To test you need to create a ServiceNow developer instance, here is my current configuration for the custom tool:

{
  "name": "ServiceNow",
  "description": "ServiceNow connector",
  "settings": {
    "base_url": "instance_url_on_portal",
    "password": "password_generated_on_portal",
    "username": "admin",
    "selected_tools": [
      "get_incidents"
    ]
  },
  "type": "servicenow"
}

Service now is facing some issues right now and my instance is down, so you may want to register a new one to try it.

https://developer.servicenow.com/dev.do